Refatoramentos
